refactored dataAccess layer

+++ storage/Models/Material.cs
@@ -2,7 +2,7 @@
 
 public class Material
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
 }
